pyproject.toml
[project]
name = "trello-ai-task-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "torch>=2.8.0",
    "transformers>=4.56.2",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
